+++ tests/src/Fixtures/MapPropertyPath/SomeAttribute.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Tests\Fixtures\MapPropertyPath;
+
+#[\Attribute(\Attribute::TARGET_METHOD | \Attribute::TARGET_PROPERTY)]
+final readonly class SomeAttribute
+{
+    public function __construct(
+        private string $value,
+    ) {}
+
+    public function getValue(): string
+    {
+        return $this->value;
+    }
+}